continue the developing  of the command part

+++ payment1/src/main/java/com/bookstore/payment1/dto/AuthorCreated.java
@@ -1,5 +1,6 @@
 package com.bookstore.payment1.dto;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 
@@ -14,13 +15,13 @@ public class AuthorCreated {
 
 	private final String name;
 
-	private final BigInteger id;
+	private final BigDecimal id;
 
 	@Singular("book")
 	private final List<CreateBook> books;
 
 	@Builder
-	public AuthorCreated(String name, List<CreateBook> books, BigInteger id) {
+	public AuthorCreated(String name, List<CreateBook> books, BigDecimal id) {
 		this.name = name;
 		this.books = books;
 		this.id = id;
