+++ payment1/src/main/java/com/bookstore/payment1/repository/AuthorRepository.java
@@ -4,12 +4,13 @@ import com.bookstore.payment1.repository.model.AuthorEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
 public interface AuthorRepository extends CrudRepository<AuthorEntity, Long> {
 
-    Optional<AuthorEntity> findByUuid(BigInteger uuid);
+    Optional<AuthorEntity> findByUuid(BigDecimal uuid);
 
 }
